Move credential management to a separate auth subcommand
