+++ doc/user/quickstart.rst
@@ -9,11 +9,11 @@ Quickstart
 Using the CLI
 -------------
 
-By default, authentication is enabled, reading in configuration settings
-from :code:`config.toml`. The path to this configuration can be found by running
-:code:`hop configure locate`. One can initialize this configuration with default
-settings by running :code:`hop configure setup`. To disable authentication in the CLI
-client, one can run :code:`--no-auth`.
+By default, authentication is enabled, reading in credentials
+from :code:`auth.toml`. The path to this configuration can be found by running
+:code:`hop auth locate`. One can initialize this configuration with default
+settings by running :code:`hop auth add`. To disable authentication in the CLI
+client, one can use the :code:`--no-auth` option.
 
 Publish messages
 ^^^^^^^^^^^^^^^^^
